Decided to alter behavior of packetpeek slightly:
packetpeek can report -1 indicating a hole in the data (not really an
error).

Previously, packetout would have to be called to clear the hole; now,
the hole is cleared by packetpeek.  That is, the change is that
ogg_stream_packetpeek must be followed by a call to
ogg_stream_packetout(state,NULL) to continue advancing forward only if
packetpeek returned a packet.


git-svn-id: http://svn.xiph.org/trunk/ogg@1454 0101bb08-14d6-0310-b084-bc0e0c8e3800